pyproject.toml
[project]
name = "1000genome-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.9.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyspark>=4.1.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
